pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "domino-cli"
dynamic = ["version"]
description = "Domino Platform CLI tool, providing utilities to install, configure and control your Domino Platform insfrastructure"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "bcrypt==4.0.0", # current latest (4.1.3) does not work
    "pyyaml",
    "requests",
]

[project.scripts]
domino-cli = "domino_cli:main"

[tool.hatch.version]
path = "domino_cli/__init__.py"

[tool.hatch.build.targets.sdist]
include = [
    "/domino_cli",
]
